made the connection with php.
Finished the connection and everything is added to the db. even edits are registered. I have to say tho, you MUST use XAMPP for how I did things. from there just create a db with the name 'task_manager' and a table with the name 'tasks' with the correct credentials. that should be it.